lambda/requirements.txt
flask==2.3.3
boto3==1.34.0
pillow==10.0.1
